Lesson 128
Finally added the Test AI nodes into the Guardian Behavior Tree and tweaked the variables there. Enemy chases when out of range, strafes around player at medium range, and there is a random chance for it to choose to approach the player. If in close range, either because player approached or enemy approached, it performs a melee attack. Motion warping effective.