Lesson 129
The default decorator to check if an actor has a tag does not inherently abort other branches, so we had to make our own new one. Also included functionality to perform an inverse check for if an actor doesn't have a specific tag.
Applied this new decorator to the Defend Self branch to check if the Self has the Enemy.Status.UnderAtttack tag and to abort lower priority branches if so.