Lesson 130
Made a new GE that adds a tag to the OwnerActor (Enemy.Status.UnderAttack). Duration policy 5 seconds. Max stack 1. Refresh on reapply. Clear stack when timer is out.
Then added the GE to the GA_HitReact_Base so it gets applied to all enemies that have hit reacts.
Set up the Defend Self BT branch to have the enemy move to target, perform a melee attack.