137 -- Roll Ability -- Two Key Input
We created 3 new Gameplay Tags:

InputTag.Roll

Player.Ability.Roll

Player.Status.Rolling

Then we create a new Gameplay Ability for Rolling. We set the tags in the details panel. The newest one is that we can set the Activation Granted tags to apply the new status tag to the Player while the ability is underway. Don’t forget to change the instancing policy to Per Actor!

We created a new IA_Roll and added the trigger Chorded Action. Under the dropdown, we selected IA_Move. This means that IA_Move will have to be triggering before we can trigger this action.

We added the new IA_Roll to the default IMC

We added elements corresponding to this new input action in both the DA_StartupConfig as well as the DA_InputConfig files.