138 -- Rolling Ability -- Directional
created the montage for rolling and played in within the GA

Don’t forget to change the slot!

So I’m excited to see where we go with this next, but it seems we will be setting up motion warping for the first few frames to adjust the character roatation based on inputs. Whoa. Cool.

Back in our gameplay ability, we create a function to compute the roll direction and distance. Important to note: inside that function we will be retrieving the Last Movement Input. Because the movement input is also triggering this selfsame ability, we cannot guarantee the order in which thoe values are fired off.

So, before we call our new function on the GA event graph, we need to add in a tiny delay of .05 secs in order to ensure the GA gets fully triggered before the function is called.