140 -- Rolling Scalable Distance by Ability level
Added in a curve to the HeroAbilities Curve table for the rolling distance so it scales at different levels. When we set this scalable float from within the GA_Roll though, we are unable to access its values.

I’m not entirely sure why, or more importantly how you would know this, but we added a helper function to our Function Library that simply takes in a scalable float and a level, and returns its value at the given level.