142 -- Direction Hit React Setup
We created the GA_Hero_HitReact and added in all the appropriate tags to cancel/block. We also need to add the Shared.Event.HitReact tag under the Trigger details panel.

In the Event Graph, we replace the Activate Ability event node with the Activate Ability from Event version since that’s where this ability will be coming from. Of course, we also make sure it is instanced per actor and not per execution.

Lastly, we generated the 4 anim montages for a hit react from different directions and changed their slot from default to UpperBody, since we will still allow the character to move even while reacting to getting hit.