143 -- HitReactActivation
Added a PlayMontageAndWait node to the GA itself. For now it will just play the FromFront montage.

Then we added the GA to the DA_HeroStartupData under Reactive Abilities

Last, we just need a way to trigger this ability on the character. Since we’ve already set it up to trigger from an event, we can send that event to the hero character from within the enemy attack GA: