145 -- Direction Hit React -- Cross Product to determine direction
We’re working in the Warrior Function Library, where we create another helper function to take in two actors, determine their angle difference, return a GameplayTag, and populate an out param that represents the angle difference.

Then we use the Cross Product to determine whether it’s coming from the right or the left. – cross product is backwards in UE, a positive result means the second vector is to the right of the first and a negative result means it is the left