146 -- Directional Hit React
We added in a bunch of Shared.Status.HitReact.XXXX tags for each hit react direction.

Now it’s time for some math. We’ve determined the angle the attack is coming from in our function library helper function, so now we just need to order that angle into a quadrant corresponding to one of the tags and send that tag as the output.