158 -- Blocking Ability
In order to handle the input correctly, we’re first going to add in a couple new tags:

InputTag_MustBeHeld

InputTag_MustBeHeld_Block

Player_Ability_Block

Player_Status_Blocking

In order to set this must be held functionality up, we need to add in code to our ASC function OnAbilityInputReleased. We first check to make sure that the ability in question doesn’t contain the MustBeHeld tag and simply return if so.

If it does contain that tag, then we will need to go through all the ability specs in the activatable abilities and find any that are active before canceling them:
This function on our ASC has a callback bound to it from the HeroCharacter.cpp, which is where it is initially called

Then we follow the usual steps:

create an IA_Block

add the tag and IA to the DA_IputConfig file

add the ability to the weapon BP abilities to grant (so the hero can only block while equipping a weapon)

add the IA and key binding to the IMX_Axe