150 -- Block -- Spawn Shield GC
Created a block anim montage and set it to loop.

Added the block tag to the canceled tags under the hit react ability.

Added the block montage to the block ability. Note: we are only calling EndAbility from the Interrupted and Canceled pins on the PlayMontageAndWait node here because the ability should stay on while the animation is looping, but should end if a strong attack breaks the block or the player manually releases the key.
when we call a GC from the GA, we have the option to do so with parameters. which we can make and pass in.

We made a Gameplay Cue, but this time the parent class is the _NotifyActor version instead of the _static version.

We override it’s WhileActive() function to play a sound and spawn the shield niagara system attached to a newly created socket on the hand_L: