153 -- Successful Block
We’re adding some more juice to our Block ability. First off, we’re rotating the character to face the enemy hit direction, second we’re applying some knockback using ApplyRootMotionConstantForce()

We made a new GameplayCue tag for successful blocks. Overriding the OnExecute function, we added in a sound and vfx particle system to spawn and gave the GC the new tag.

Within the GA, we called ExecuteGameplayCueWithParamsOnOwner and passed in the newly created tag (GameplayCue.Effects.MagicShield.SuccessfulBlock). For the params passed in, we are just making that struct and passing in the skeletal mesh component so the GC can access the MagicShieldSocket on it for where the particle should spawn.