154 -- Perfect Block
This one is easier than I expected. We will get the time in seconds when the ability is activated and cache that value in a variable. Then, right after the WaitGameplayEvent triggers, we can check the tine in seconds again, and subtract the cached value. If it is less than a certain threshold, that means the player activated block right before the hit actually came in, and it should be treated as a perfect block instead of just a normal block

 NOTE/TODO: the threshold value can be hardcoded like we have it now, but it can also be tied to a curve table – allowing perfect blocks to get easier as the player abilities level up!