156 Counterattack
In our Block ability, if the player achieves a perfect block, the first thing we’re now doing is adding the JumpToFinisher tag we made previously to the player. So their counterattack ability is allowing them to perform a finisher attack.

We also have to handle removing the tag if they do not perform an attack quickly enough, so we set a timer by event that removes the tag. We call the start timer function right after the time dilation and anywhere else that EndAbility is called.
Our Heavy attack ability already has functionality for checking the JumpToFinisher tag, but we need to add it into our Light Attack ability. Just check to see if actor has tag, and if so, set the combo count to the length of the montages map.