158 -- TargetLock Setup
The first thing we’re doing for target lock is to create a new GA in c++ derived from our HeroGameplayAbility. We override the native functions ActivateAbility and EndAbility. Here are all the “Important Functions” within the base class:
We added in a bunch of tags for this ability, including InputTag tags for toggleable and toggleable_TargetLock, an ability tag and a status tag.

When we make our GA in blueprints, one major difference is that we are leaving the instancing policy on once per execution because we do not want to save variables between activations – we want a clean slate each time.

Otherwise, we just give it the newly created ability tag and the status tag (under activation given).