159 -- Toggle TargetLock
In order to handle a toggleable ability, we need to go into our ASC and adjust how ability input activations are being handled. First of all, we need to check if the ability has the toggleable tag, then turn it off if it is already active, otherwise turn it on.

I just noticed the code above is wrong, we need to keep another line for TryActivateAbility() outside the if check for whether it is toggleable, otherwise our normal abilities won’t fire.

added input action and added it to Axe IMC
granted ability through the heroweapon_axe
updated input config