160 -- Get all avalaible targets for lock
Ok wow. We’re doing a lot here, and I’m a little curious about the UKismetSystemLibrary that we’re using to perform a BoxMultiTraceForObjects(). I think I understand everything that is going into it as a parameter, but I’m a little confused on how exactly it does what it does. I guess we’re just sending out a huge box in front of us? And hoping that collects every target? seems somewhat wasteful but whatever.

Then, we’re collecting an array of possible targets by looping through everything we hit and adding them in.