161 -- Find the nearest actor
Lots of new code in this one.

We added in a handful of functions to help us filter through the BoxTrace hitresults and return the closest valid actor. As well as a handful of variables editable from the details panel.

Worth noting is that GameplayStatics has a FindNearestActor() func that we can use rather than reinventing the wheel.