162 -- TargetLock Widget
First thing, we created a very simple WBP TargetLock widget derived from our custom widget class and using our custom sizebox class – I can’t remember why we created those, but they’re available which is great.

When using the CreateWidget func, it throws an error when we try to assign its owner as the customer Hero Controller class, so we need to include that class’s header file. Not an obvious error:
And even once the .h file is included, Rider thinks it is not being used and greys it out, yet the error is gone. Unintuitive to the max.