163 Setting Widget Position
So this is thei SetTargetLockWidgetPosition() function that we call in TryLockOnTarget() after the widget is drawn and added to the viewport.

first we’re checking to make sure the current locked actor and drawn widget are both valid before moving on to Projecting the World Location of the actor to a widget position. This is a function within the UWidgetLayoutLibrary that esstentially translates a Vector 3D to a Vector2D.

Then, if the Vector2D variable we have for the widget size hasn’t been initiated, we initiate it based on the sizebox in the widget. Useful little function there eh?

Then we take that size box’s half size and subtract it from the screen position variable we calculated before we actually draw and put it in viewport.

Next, we’ll work on updating the widget position on tick so it isn’t stuck in a single screen pos on initialization.