165-- Update Widget Position
Now that we have tick being registered in the GA_TargetLock through the use of our custom AbilityTask, the next thing to do is update the widget’s position on tick.

For that, we’ll create a new function within the cpp version of the GA called OnTargetLockTick.

We first check to make sure the CurrentLockedActor is valid and neither it nor the hero character have the Status.Dead tag. If that check passes we simply call the SetWidgetPosition function