166 Orient the char
In our newly created function that draws the widget to the screen on each tick, we also need to add in functionality to orient the hero character. This includes both the Avatar actor as well as the camera. Luckily, the camera is already set to follow the controller rotation, so if we set that directly we should be good.

For this, we will first check to make sure the character is not rolling or blocking, in which case, those animations should drive the character rotation (although maybe if the character is locked on and blocking, they should follow the target?)

As long as neither of those is applicable, we will make a LookAt rotator based on the character location and the target locked actor location. This will be the target rotation to interp to. Then we get the controller rotation and calculate the interp to rotation as our target rotation for that particular Tick.

Lastly, we set the controller rotation to the new target’s pitch and yaw (but not roll) and we set the character rotation to the new target’s yaw (but not pitch or roll)