167 -- TargetLock Blendspace
In this one we made a blendspace to handle the character moving in different directions while target locked onto an enemy. The bottom nodes are all idle pose, while the top nodes are all either fwd/left/right/bwd walking animations, with a little fix for the r/l transitional states.