Switching Target Lock
Created a new IA for SwitchTarget.

Bound it to the controller as a native Input

Added it to the Input_Config data.

Created a new IMC for Switch Target with the same input binding as IA_Look

Added functions into the GA to add/remove this IMC with a higher priority than the default so it overrides IA Look.

Added a variable in the GA to set the new IMC.
When removing the IMC, don’t forget to check() the PlayerController first, otherwise it may try removing it from a null pointer if exiting game while the ability is active