201 & 202
set up hit react ability for the boss. this time it doesn't cancel or block the abilities and doesn't adjust the rotation of the boss

also added in the death ability. created the montage and gameplay cue sound notify with tag. the niagara dissolve was already created.

next we need to resolve the boss health bar after death.